Commit One (Adding the Game)
This is a java project done using processing.

This is a simple game where the player is able to click on the circles and pop the circles like a balloon.

Additionally, there is a scoring system that currently does not work correctly.

Secondly, there is a section for a second game if the player presses the number 2 key there is a section where a second game can be introduced however currently there is not a second game in the second section. The hope is to add more games in the future.